requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Vision Transformer
timm>=0.9.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# Machine Learning
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.23.0
